fix(providers): repair openai compat CI regressions
