+++ crates/providers/src/openai_compat/mod.rs
@@ -4,6 +4,10 @@
 //! that include tool calls. Used by openai.rs, github_copilot.rs, and kimi_code.rs.
 
 mod schema_normalization;
+mod strict_mode;
+
+#[cfg(test)]
+mod tests;
 
 use std::collections::{HashMap, HashSet};
 
@@ -13,6 +17,8 @@ use moltis_agents::model::{
     ChatMessage, CompletionResponse, StreamEvent, ToolCall, Usage, UserContent,
     decode_tool_call_arguments,
 };
+
+pub use strict_mode::patch_schema_for_strict_mode;
 
 // ============================================================================
 // OpenAI Tool Schema Types
@@ -59,182 +65,6 @@ pub struct ResponsesApiTool {
     pub description: String,
     pub parameters: serde_json::Value,
     pub strict: bool,
-}
-
-/// Make a JSON Schema node nullable by adding `"null"` to its type.
-///
-/// For schemas using `"type": "string"` (or any single type), this converts to
-/// `"type": ["string", "null"]`.  For schemas using `anyOf`/`oneOf`, it appends
-/// a `{"type": "null"}` variant.  Already-nullable schemas are left unchanged.
-fn make_nullable(schema: &mut serde_json::Value) {
-    let Some(obj) = schema.as_object_mut() else {
-        return;
-    };
-
-    // If it already has a "null" type, nothing to do.
-    if let Some(ty) = obj.get("type") {
-        if ty.as_str() == Some("null") {
-            return;
-        }
-        if let Some(arr) = ty.as_array()
-            && arr.iter().any(|v| v.as_str() == Some("null"))
-        {
-            return;
-        }
-    }
-
-    // Case 1: has a `type` field — convert to array form with "null".
-    if let Some(ty) = obj.get("type").cloned() {
-        if let Some(s) = ty.as_str() {
-            obj.insert("type".to_string(), serde_json::json!([s, "null"]));
-        } else if let Some(arr) = ty.as_array() {
-            let mut new_arr = arr.clone();
-            new_arr.push(serde_json::json!("null"));
-            obj.insert("type".to_string(), serde_json::Value::Array(new_arr));
-        }
-        return;
-    }
-
-    // Case 2: uses anyOf/oneOf — append a null variant.
-    for key in ["anyOf", "oneOf"] {
-        if let Some(variants) = obj.get_mut(key).and_then(|v| v.as_array_mut()) {
-            let has_null = variants
-                .iter()
-                .any(|v| v.get("type").and_then(|t| t.as_str()) == Some("null"));
-            if !has_null {
-                variants.push(serde_json::json!({"type": "null"}));
-            }
-            return;
-        }
-    }
-}
-
-/// Check whether a schema's `type` field designates an object.
-///
-/// Matches both `"type": "object"` and `"type": ["object", ...]`.
-/// Downstream providers like Gemini (via OpenRouter) cannot represent type
-/// arrays, and OpenAI's strict mode requires `type` to be a single string.
-fn type_is_object(obj: &serde_json::Map<String, serde_json::Value>) -> bool {
-    let Some(ty) = obj.get("type") else {
-        return false;
-    };
-    match ty {
-        serde_json::Value::String(s) => s == "object",
-        serde_json::Value::Array(arr) => arr.iter().any(|v| v.as_str() == Some("object")),
-        _ => false,
-    }
-}
-
-/// Recursively patch schema for OpenAI strict mode compliance.
-///
-/// OpenAI's strict mode requires:
-/// 1. `additionalProperties: false` on every object in the schema tree
-/// 2. All properties must be listed in the `required` array
-///
-/// Properties not in the original `required` array are made nullable so the
-/// model can send `null` for unused optional fields instead of fabricating
-/// placeholder values.
-///
-/// Schemas with `type: ["object", "string", ...]` (union types) are collapsed
-/// to `type: "object"` because:
-/// - OpenAI strict mode requires `type` to be a single string
-/// - Gemini cannot represent type arrays at all
-/// - When `properties` is present, the object form is the intended shape for
-///   tool calling (bare strings/integers are convenience shorthand for non-tool
-///   contexts)
-///
-/// This function recursively patches nested objects in `properties`, array
-/// `items`, `anyOf`/`oneOf`/`allOf` variants, etc.
-pub fn patch_schema_for_strict_mode(schema: &mut serde_json::Value) {
-    let Some(obj) = schema.as_object_mut() else {
-        return;
-    };
-
-    // Collect originally-optional property names so we can make them nullable
-    // AFTER recursion (otherwise changing "type":"object" to ["object","null"]
-    // would prevent the recursive pass from recognising nested objects).
-    let mut optional_props: Vec<String> = Vec::new();
-
-    // If this is an object type, apply strict mode requirements.
-    // Also handles union types like ["object", "string"] by collapsing to "object".
-    if type_is_object(obj) {
-        // Collapse type arrays (e.g. ["object", "string"]) to "object".
-        // OpenAI strict mode and Gemini both require a single string type.
-        if obj.get("type").and_then(|t| t.as_array()).is_some() {
-            obj.insert("type".to_string(), serde_json::json!("object"));
-        }
-        // Add additionalProperties: false
-        obj.insert("additionalProperties".to_string(), serde_json::json!(false));
-
-        // Ensure all properties are in required array.
-        if let Some(props) = obj.get("properties").and_then(|p| p.as_object()).cloned() {
-            let originally_required: HashSet<String> = obj
-                .get("required")
-                .and_then(|r| r.as_array())
-                .map(|arr| {
-                    arr.iter()
-                        .filter_map(|v| v.as_str().map(String::from))
-                        .collect()
-                })
-                .unwrap_or_default();
-
-            let all_prop_names: Vec<serde_json::Value> =
-                props.keys().map(|k| serde_json::json!(k)).collect();
-            obj.insert("required".to_string(), serde_json::json!(all_prop_names));
-
-            // Remember which properties need to be made nullable (deferred).
-            for key in props.keys() {
-                if !originally_required.contains(key) {
-                    optional_props.push(key.clone());
-                }
-            }
-        } else {
-            // Object without properties - add empty properties and required
-            obj.insert("properties".to_string(), serde_json::json!({}));
-            obj.insert("required".to_string(), serde_json::json!([]));
-        }
-    }
-
-    // Recurse into properties
-    if let Some(props) = obj.get_mut("properties").and_then(|p| p.as_object_mut()) {
-        for (_, prop_schema) in props.iter_mut() {
-            patch_schema_for_strict_mode(prop_schema);
-        }
-    }
-
-    // Recurse into array items
-    if let Some(items) = obj.get_mut("items") {
-        patch_schema_for_strict_mode(items);
-    }
-
-    // Recurse into anyOf/oneOf/allOf
-    for key in ["anyOf", "oneOf", "allOf"] {
-        if let Some(variants) = obj.get_mut(key).and_then(|v| v.as_array_mut()) {
-            for variant in variants {
-                patch_schema_for_strict_mode(variant);
-            }
-        }
-    }
-
-    // Recurse into additionalProperties if it's a schema (not just true/false)
-    if let Some(additional) = obj.get_mut("additionalProperties")
-        && additional.is_object()
-    {
-        patch_schema_for_strict_mode(additional);
-    }
-
-    // Now that recursion is done, make originally-optional properties nullable
-    // so the model can send `null` instead of fabricating placeholder values
-    // (e.g. empty strings) that downstream MCP servers reject.
-    if !optional_props.is_empty()
-        && let Some(props) = obj.get_mut("properties").and_then(|p| p.as_object_mut())
-    {
-        for key in &optional_props {
-            if let Some(prop_schema) = props.get_mut(key) {
-                make_nullable(prop_schema);
-            }
-        }
-    }
 }
 
 /// Convert tool schemas to OpenAI Chat Completions function-calling format.
@@ -1143,422 +973,5 @@ pub fn parse_responses_completion(resp: &serde_json::Value) -> CompletionRespons
         text,
         tool_calls,
         usage,
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{
-        parse_responses_completion, parse_tool_calls, sanitize_schema_for_openai_compat,
-        to_openai_tools, to_responses_api_tools,
-    };
-
-    #[test]
-    fn parse_tool_calls_preserves_native_falsy_types() {
-        let msg = serde_json::json!({
-            "tool_calls": [{
-                "id": "call_1",
-                "function": {
-                    "name": "grep",
-                    "arguments": {
-                        "offset": 0,
-                        "multiline": false,
-                        "type": null
-                    }
-                }
-            }]
-        });
-
-        let calls = parse_tool_calls(&msg);
-
-        assert_eq!(calls.len(), 1);
-        assert_eq!(calls[0].arguments["offset"], 0);
-        assert_eq!(calls[0].arguments["multiline"], false);
-        assert!(calls[0].arguments["type"].is_null());
-    }
-
-    #[test]
-    fn parse_tool_calls_preserve_issue_693_examples() {
-        let msg = serde_json::json!({
-            "tool_calls": [
-                {
-                    "id": "call_exec",
-                    "function": {
-                        "name": "exec",
-                        "arguments": {
-                            "command": "echo hello",
-                            "timeout": 0
-                        }
-                    }
-                },
-                {
-                    "id": "call_edit",
-                    "function": {
-                        "name": "Edit",
-                        "arguments": {
-                            "replace_all": false
-                        }
-                    }
-                },
-                {
-                    "id": "call_grep",
-                    "function": {
-                        "name": "Grep",
-                        "arguments": {
-                            "offset": 0,
-                            "multiline": false,
-                            "type": null
-                        }
-                    }
-                }
-            ]
-        });
-
-        let calls = parse_tool_calls(&msg);
-
-        assert_eq!(calls.len(), 3);
-        assert_eq!(calls[0].arguments["timeout"], 0);
-        assert_eq!(calls[1].arguments["replace_all"], false);
-        assert_eq!(calls[2].arguments["offset"], 0);
-        assert_eq!(calls[2].arguments["multiline"], false);
-        assert!(calls[2].arguments["type"].is_null());
-    }
-
-    #[test]
-    fn parse_responses_completion_preserves_native_falsy_types() {
-        let resp = serde_json::json!({
-            "output": [{
-                "type": "function_call",
-                "call_id": "call_abc",
-                "name": "grep",
-                "arguments": {
-                    "offset": 0,
-                    "multiline": false,
-                    "type": null
-                }
-            }],
-            "usage": {"input_tokens": 20, "output_tokens": 10}
-        });
-
-        let result = parse_responses_completion(&resp);
-
-        assert_eq!(result.tool_calls.len(), 1);
-        assert_eq!(result.tool_calls[0].arguments["offset"], 0);
-        assert_eq!(result.tool_calls[0].arguments["multiline"], false);
-        assert!(result.tool_calls[0].arguments["type"].is_null());
-    }
-
-    #[test]
-    fn responses_tools_strip_nested_not_schemas() {
-        let tools = vec![serde_json::json!({
-            "name": "mcp__attio__list-attribute-definitions",
-            "description": "Attio test tool",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "query": {
-                        "anyOf": [
-                            {
-                                "anyOf": [
-                                    {
-                                        "not": {
-                                            "const": ""
-                                        }
-                                    },
-                                    {
-                                        "type": "string"
-                                    }
-                                ]
-                            },
-                            {
-                                "type": "null"
-                            }
-                        ]
-                    }
-                }
-            }
-        })];
-
-        let converted = to_responses_api_tools(&tools);
-        let params = &converted[0]["parameters"];
-        let encoded = params.to_string();
-
-        assert_eq!(converted[0]["strict"], true);
-        assert!(!encoded.contains("\"not\""));
-        assert_eq!(params["type"], "object");
-        assert_eq!(params["additionalProperties"], false);
-        assert_eq!(params["required"], serde_json::json!(["query"]));
-    }
-
-    #[test]
-    fn sanitize_schema_for_openai_compat_strips_recursive_unsupported_keywords() {
-        let mut schema = serde_json::json!({
-            "type": "object",
-            "properties": {
-                "config": {
-                    "type": "object",
-                    "properties": {
-                        "mode": { "type": "string" }
-                    },
-                    "if": {
-                        "required": ["mode"]
-                    },
-                    "then": {
-                        "properties": {
-                            "enabled": { "type": "boolean" }
-                        }
-                    },
-                    "else": {
-                        "properties": {
-                            "enabled": { "type": "boolean" }
-                        }
-                    },
-                    "dependentSchemas": {
-                        "mode": {
-                            "properties": {
-                                "extra": { "type": "string" }
-                            }
-                        }
-                    },
-                    "patternProperties": {
-                        "^x-": { "type": "string" }
-                    },
-                    "dependentRequired": {
-                        "mode": ["enabled"]
-                    },
-                    "unevaluatedProperties": false,
-                    "unevaluatedItems": false,
-                    "propertyNames": {
-                        "minLength": 1
-                    },
-                    "contains": {
-                        "type": "string"
-                    },
-                    "minContains": 1,
-                    "maxContains": 2,
-                    "minProperties": 1,
-                    "maxProperties": 4,
-                    "const": "active",
-                    "x-custom": "remove-me",
-                    "items": {
-                        "not": {
-                            "type": "integer"
-                        }
-                    }
-                }
-            }
-        });
-
-        sanitize_schema_for_openai_compat(&mut schema);
-        let encoded = schema.to_string();
-
-        for keyword in [
-            "\"if\"",
-            "\"then\"",
-            "\"else\"",
-            "\"dependentSchemas\"",
-            "\"patternProperties\"",
-            "\"dependentRequired\"",
-            "\"unevaluatedProperties\"",
-            "\"unevaluatedItems\"",
-            "\"propertyNames\"",
-            "\"contains\"",
-            "\"minContains\"",
-            "\"maxContains\"",
-            "\"minProperties\"",
-            "\"maxProperties\"",
-            "\"not\"",
-            "\"x-custom\"",
-        ] {
-            assert!(!encoded.contains(keyword), "{keyword} should be removed");
-        }
-        assert_eq!(
-            schema["properties"]["config"]["enum"],
-            serde_json::json!(["active"])
-        );
-        assert_eq!(
-            schema["properties"]["config"]["properties"]["mode"]["type"],
-            "string"
-        );
-    }
-
-    #[test]
-    fn sanitize_schema_for_openai_compat_recurses_into_array_form_items() {
-        let mut schema = serde_json::json!({
-            "type": "object",
-            "properties": {
-                "tuple": {
-                    "type": "array",
-                    "items": [
-                        {
-                            "type": "string",
-                            "not": { "const": "" }
-                        },
-                        {
-                            "type": "object",
-                            "patternProperties": {
-                                "^x-": { "type": "string" }
-                            }
-                        }
-                    ]
-                }
-            }
-        });
-
-        sanitize_schema_for_openai_compat(&mut schema);
-
-        let Some(tuple_items) = schema["properties"]["tuple"]["items"].as_array() else {
-            panic!("tuple items should remain an array");
-        };
-        assert!(tuple_items[0].get("not").is_none());
-        assert!(tuple_items[1].get("patternProperties").is_none());
-    }
-
-    #[test]
-    fn to_openai_tools_strict_mode_applied_by_default() {
-        let tools = vec![serde_json::json!({
-            "name": "create_file",
-            "description": "Create a file",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "path": { "type": "string" },
-                    "content": { "type": "string" },
-                    "overwrite": { "type": "boolean" }
-                },
-                "required": ["path"]
-            }
-        })];
-
-        let converted = to_openai_tools(&tools, true);
-        assert_eq!(converted.len(), 1);
-
-        let func = &converted[0]["function"];
-        assert_eq!(func["strict"], true);
-        assert_eq!(func["parameters"]["additionalProperties"], false);
-
-        // All properties must be in the required array.
-        let Some(required) = func["parameters"]["required"].as_array() else {
-            panic!("required should be an array");
-        };
-        let required_names: Vec<&str> = required.iter().filter_map(|v| v.as_str()).collect();
-        assert!(required_names.contains(&"path"));
-        assert!(required_names.contains(&"content"));
-        assert!(required_names.contains(&"overwrite"));
-    }
-
-    #[test]
-    fn to_openai_tools_non_strict_skips_patching() {
-        let tools = vec![serde_json::json!({
-            "name": "create_file",
-            "description": "Create a file",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "path": { "type": "string" },
-                    "content": { "type": "string" },
-                    "overwrite": { "type": "boolean" }
-                },
-                "required": ["path"]
-            }
-        })];
-
-        let converted = to_openai_tools(&tools, false);
-        assert_eq!(converted.len(), 1);
-
-        let func = &converted[0]["function"];
-        assert_eq!(func["strict"], false);
-
-        // strict-mode patching must NOT have run: no additionalProperties
-        // and no array-form nullable types like ["boolean","null"].
-        let serialized = func["parameters"].to_string();
-        assert!(
-            !serialized.contains("additionalProperties"),
-            "strict mode should not inject additionalProperties: {serialized}"
-        );
-        assert!(
-            !serialized.contains("[\"boolean\""),
-            "strict mode should not produce array-form types: {serialized}"
-        );
-        assert!(
-            !serialized.contains("[\"string\""),
-            "strict mode should not produce array-form types: {serialized}"
-        );
-    }
-
-    #[test]
-    fn to_openai_tools_non_strict_complex_cron_like_schema() {
-        // Schema modeled after the cron tool's `job` property: nested objects,
-        // optional booleans, enum fields — the shapes that trigger Google's
-        // INVALID_ARGUMENT when strict-mode array types are present.
-        let tools = vec![serde_json::json!({
-            "name": "schedule_cron",
-            "description": "Schedule a cron job",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "name": { "type": "string" },
-                    "job": {
-                        "type": "object",
-                        "properties": {
-                            "enabled": { "type": "boolean" },
-                            "schedule": { "type": "string" },
-                            "retry": { "type": "boolean" },
-                            "mode": {
-                                "type": "string",
-                                "enum": ["once", "recurring"]
-                            },
-                            "config": {
-                                "type": "object",
-                                "properties": {
-                                    "timeout": { "type": "integer" },
-                                    "verbose": { "type": "boolean" }
-                                },
-                                "required": ["timeout"]
-                            }
-                        },
-                        "required": ["schedule"]
-                    }
-                },
-                "required": ["name", "job"]
-            }
-        })];
-
-        let converted = to_openai_tools(&tools, false);
-        let func = &converted[0]["function"];
-        assert_eq!(func["strict"], false);
-
-        // No array-form types anywhere in the output.
-        let serialized = func["parameters"].to_string();
-        // Array-form types look like `["boolean","null"]` — must not appear.
-        assert!(
-            !serialized.contains("[\"boolean\""),
-            "should not contain array-form types: {serialized}"
-        );
-        assert!(
-            !serialized.contains("[\"string\""),
-            "should not contain array-form types: {serialized}"
-        );
-        assert!(
-            !serialized.contains("[\"integer\""),
-            "should not contain array-form types: {serialized}"
-        );
-
-        // Original required arrays preserved.
-        let Some(job_required) = func["parameters"]["properties"]["job"]["required"].as_array()
-        else {
-            panic!("job required should be an array");
-        };
-        assert_eq!(job_required.len(), 1);
-        assert_eq!(job_required[0], "schedule");
-
-        // Nested object required array also preserved.
-        let Some(config_required) =
-            func["parameters"]["properties"]["job"]["properties"]["config"]["required"].as_array()
-        else {
-            panic!("config required should be an array");
-        };
-        assert_eq!(config_required.len(), 1);
-        assert_eq!(config_required[0], "timeout");
     }
 }
