+++ crates/providers/tests/openai_compat_type_arrays.rs
@@ -1,17 +1,7 @@
-/// Tests for type array collapsing in strict mode patching.
-///
-/// When tool schemas have `type: ["object", "string", ...]`, the strict mode
-/// patcher must collapse these to `type: "object"` because:
-/// - OpenAI strict mode requires `type` to be a single string
-/// - Gemini cannot represent type arrays at all
-/// - The "object" form is the intended shape for tool calling
-///
-/// See: https://github.com/moltis-org/moltis/issues/716
 use moltis_providers::openai_compat::{patch_schema_for_strict_mode, to_openai_tools};
 
 #[test]
 fn strict_mode_collapses_object_union_types() {
-    // type: ["object", "string"] should collapse to "object"
     let mut schema = serde_json::json!({
         "type": ["object", "string"],
         "properties": {
@@ -31,7 +21,6 @@ fn strict_mode_collapses_object_union_types() {
 
 #[test]
 fn strict_mode_collapses_triple_union_types() {
-    // type: ["object", "string", "integer"] should collapse to "object"
     let mut schema = serde_json::json!({
         "type": ["object", "string", "integer"],
         "properties": {
@@ -61,7 +50,6 @@ fn strict_mode_leaves_singular_string_type_unchanged() {
 
 #[test]
 fn strict_mode_leaves_non_object_array_unchanged() {
-    // type: ["string", "integer"] — no "object", no collapse
     let mut schema = serde_json::json!({
         "type": ["string", "integer"]
     });
@@ -99,29 +87,23 @@ fn strict_mode_collapses_nested_object_union_types() {
 
     patch_schema_for_strict_mode(&mut schema);
 
-    // Top level: singular "object" — unchanged type but patched
     assert_eq!(schema["type"], "object");
-
-    // schedule: collapsed from array to "object"
     assert_eq!(schema["properties"]["schedule"]["type"], "object");
     assert_eq!(
         schema["properties"]["schedule"]["additionalProperties"],
         false
     );
-
-    // payload: collapsed from array to "object"
     assert_eq!(schema["properties"]["payload"]["type"], "object");
     assert_eq!(
         schema["properties"]["payload"]["additionalProperties"],
         false
     );
 
-    // All required entries should match actual properties
     let top_required: Vec<&str> = schema["required"]
         .as_array()
         .unwrap()
         .iter()
-        .map(|v| v.as_str().unwrap())
+        .map(|value| value.as_str().unwrap())
         .collect();
     for name in &top_required {
         assert!(
@@ -162,12 +144,10 @@ fn to_openai_tools_collapses_union_types_end_to_end() {
         }
     })];
 
-    let converted = to_openai_tools(&tools);
+    let converted = to_openai_tools(&tools, true);
     assert_eq!(converted.len(), 1);
 
     let params = &converted[0]["function"]["parameters"];
-
-    // No type arrays should remain
     let mut type_arrays = Vec::new();
     find_type_arrays(params, "root", &mut type_arrays);
     assert!(
@@ -175,17 +155,13 @@ fn to_openai_tools_collapses_union_types_end_to_end() {
         "found type arrays after pipeline: {type_arrays:?}"
     );
 
-    // All required entries should exist in properties
     let mut orphans = Vec::new();
     find_required_orphans(params, "root", &mut orphans);
     assert!(orphans.is_empty(), "found required orphans: {orphans:?}");
 }
 
-// --- Additional coverage for recursion paths and edge cases ---
-
 #[test]
-fn collapse_inside_anyOf_variants() {
-    // Type arrays inside anyOf variants should be collapsed recursively.
+fn collapse_inside_any_of_variants() {
     let mut schema = serde_json::json!({
         "type": "object",
         "properties": {
@@ -202,15 +178,13 @@ fn collapse_inside_anyOf_variants() {
     patch_schema_for_strict_mode(&mut schema);
 
     let any_of = schema["properties"]["config"]["anyOf"].as_array().unwrap();
-    // First variant: collapsed from ["object", "string"] to "object"
     assert_eq!(any_of[0]["type"], "object");
     assert_eq!(any_of[0]["additionalProperties"], false);
-    // Second variant: plain string, untouched
     assert_eq!(any_of[1]["type"], "string");
 }
 
 #[test]
-fn collapse_inside_oneOf_variants() {
+fn collapse_inside_one_of_variants() {
     let mut schema = serde_json::json!({
         "type": "object",
         "properties": {
@@ -235,7 +209,6 @@ fn collapse_inside_oneOf_variants() {
 
 #[test]
 fn collapse_inside_array_items() {
-    // Type array inside array items should be collapsed.
     let mut schema = serde_json::json!({
         "type": "object",
         "properties": {
@@ -260,8 +233,6 @@ fn collapse_inside_array_items() {
 
 #[test]
 fn collapsed_optional_object_becomes_nullable() {
-    // A union-type object that is NOT in the original required array
-    // should become ["object", "null"] after strict mode patching.
     let mut schema = serde_json::json!({
         "type": "object",
         "properties": {
@@ -277,24 +248,16 @@ fn collapsed_optional_object_becomes_nullable() {
 
     patch_schema_for_strict_mode(&mut schema);
 
-    // "config" was optional, so it should now be nullable
     let config_type = &schema["properties"]["config"]["type"];
     let arr = config_type
         .as_array()
         .expect("optional union-type object should have array type after nullable conversion");
-    assert!(
-        arr.contains(&serde_json::json!("object")),
-        "type should contain 'object'"
-    );
-    assert!(
-        arr.contains(&serde_json::json!("null")),
-        "type should contain 'null' for optional property"
-    );
+    assert!(arr.contains(&serde_json::json!("object")));
+    assert!(arr.contains(&serde_json::json!("null")));
 }
 
 #[test]
 fn object_not_first_in_array_still_collapses() {
-    // "object" can appear anywhere in the array, not just first.
     let mut schema = serde_json::json!({
         "type": ["string", "integer", "object"],
         "properties": { "id": { "type": "string" } },
@@ -309,7 +272,6 @@ fn object_not_first_in_array_still_collapses() {
 
 #[test]
 fn empty_type_array_is_not_object() {
-    // type: [] — no "object" present, should not be treated as object.
     let mut schema = serde_json::json!({
         "type": []
     });
@@ -320,24 +282,23 @@ fn empty_type_array_is_not_object() {
     assert!(schema.get("additionalProperties").is_none());
 }
 
-// --- Test helpers ---
-
 fn find_type_arrays(schema: &serde_json::Value, path: &str, results: &mut Vec<String>) {
     let Some(obj) = schema.as_object() else {
         return;
     };
-    if let Some(arr) = obj.get("type").and_then(|t| t.as_array()) {
+
+    if let Some(arr) = obj.get("type").and_then(|value| value.as_array()) {
         results.push(format!("{path}: type={arr:?}"));
     }
-    if let Some(props) = obj.get("properties").and_then(|p| p.as_object()) {
-        for (key, val) in props {
-            find_type_arrays(val, &format!("{path}.{key}"), results);
+    if let Some(props) = obj.get("properties").and_then(|value| value.as_object()) {
+        for (key, value) in props {
+            find_type_arrays(value, &format!("{path}.{key}"), results);
         }
     }
     for key in ["anyOf", "oneOf", "allOf"] {
-        if let Some(variants) = obj.get(key).and_then(|v| v.as_array()) {
-            for (i, variant) in variants.iter().enumerate() {
-                find_type_arrays(variant, &format!("{path}.{key}[{i}]"), results);
+        if let Some(variants) = obj.get(key).and_then(|value| value.as_array()) {
+            for (index, variant) in variants.iter().enumerate() {
+                find_type_arrays(variant, &format!("{path}.{key}[{index}]"), results);
             }
         }
     }
@@ -350,32 +311,32 @@ fn find_required_orphans(schema: &serde_json::Value, path: &str, results: &mut V
     let Some(obj) = schema.as_object() else {
         return;
     };
+
     if let (Some(required), Some(properties)) = (
-        obj.get("required").and_then(|r| r.as_array()),
-        obj.get("properties").and_then(|p| p.as_object()),
+        obj.get("required").and_then(|value| value.as_array()),
+        obj.get("properties").and_then(|value| value.as_object()),
     ) {
         for entry in required {
-            if let Some(name) = entry.as_str() {
-                if !properties.contains_key(name) {
-                    results.push(format!("{path}: required '{name}' not in properties"));
-                }
+            if let Some(name) = entry.as_str()
+                && !properties.contains_key(name)
+            {
+                results.push(format!("{path}: required '{name}' not in properties"));
             }
         }
     }
-    if let Some(props) = obj.get("properties").and_then(|p| p.as_object()) {
-        for (key, val) in props {
-            find_required_orphans(val, &format!("{path}.{key}"), results);
+    if let Some(props) = obj.get("properties").and_then(|value| value.as_object()) {
+        for (key, value) in props {
+            find_required_orphans(value, &format!("{path}.{key}"), results);
         }
     }
     for key in ["anyOf", "oneOf", "allOf"] {
-        if let Some(variants) = obj.get(key).and_then(|v| v.as_array()) {
-            for (i, variant) in variants.iter().enumerate() {
-                find_required_orphans(variant, &format!("{path}.{key}[{i}]"), results);
+        if let Some(variants) = obj.get(key).and_then(|value| value.as_array()) {
+            for (index, variant) in variants.iter().enumerate() {
+                find_required_orphans(variant, &format!("{path}.{key}[{index}]"), results);
             }
         }
     }
     if let Some(items) = obj.get("items") {
         find_required_orphans(items, &format!("{path}.items"), results);
     }
-}
 }
